fixed comments from last PR
