+++ src/app/modules/game/defeat-modal/defeat-modal.component.ts
@@ -2,17 +2,17 @@ import { Component } from '@angular/core';
 import { NgbActiveModal } from '@ng-bootstrap/ng-bootstrap';
 
 @Component({
-  selector: 'app-quit-game-confirm',
+  selector: 'app-defeat-modal',
   template: `
     <div class="modal-content">
       <div class="modal-header ">
-        <h4 class="modal-title w-100 text-center">Quitter</h4>
+        <h4 class="modal-title w-100 text-center">Défaite</h4>
         <button type="button" class="close" aria-label="Close" (click)="dismiss()">
           <span aria-hidden="true">&times;</span>
         </button>
       </div>
       <div class="modal-body w-100 text-center">
-        <p>Êtes vous sûr de vouloir quitter la partie en cours ?</p>
+        <p>Voulez vous rejouer ?</p>
       </div>
       <div class="modal-footer justify-content-center">
         <button type="button" class="btn-yes" (click)="close()">Oui</button>
@@ -22,7 +22,7 @@ import { NgbActiveModal } from '@ng-bootstrap/ng-bootstrap';
   `,
   styleUrls: ['../../shared/modals.scss'],
 })
-export class QuitGameConfirmationModalComponent {
+export class DefeatModalComponent {
   constructor(private modal: NgbActiveModal) {}
 
   dismiss() {
