+++ src/app/modules/game/game.module.ts
@@ -13,6 +13,9 @@ import { QuestionPointsComponent } from 'src/app/modules/game/question-points/qu
 import { QuestionComponent } from 'src/app/modules/game/question/question.component';
 import { TimerComponent } from 'src/app/modules/game/timer/timer.component';
 import { SharedModule } from 'src/app/modules/shared/shared.module';
+import { EndOfTurnModalComponent } from './end-of-turn-modal/end-of-turn-modal.component';
+import { VictoryModalComponent } from './victory-modal/victory-modal.component';
+import { DefeatModalComponent } from './defeat-modal/defeat-modal.component';
 
 @NgModule({
   declarations: [
@@ -26,7 +29,12 @@ import { SharedModule } from 'src/app/modules/shared/shared.module';
     CriterionPointsComponent,
     QuestionPointsComponent,
     TimerComponent,
+    EndOfTurnModalComponent,
+    VictoryModalComponent,
+    DefeatModalComponent,
   ],
+  entryComponents: [EndOfTurnModalComponent, VictoryModalComponent, DefeatModalComponent],
   imports: [CommonModule, LuxonModule, SharedModule, GameRoutingModule],
+  exports: [EndOfTurnModalComponent, VictoryModalComponent, DefeatModalComponent],
 })
 export class GameModule {}
