+++ src/app/modules/game/victory-modal/victory-modal.component.ts
