+++ src/app/modules/pregame/missing-level-selection-modal/missing-level-selection-modal.component.ts
@@ -2,27 +2,26 @@ import { Component } from '@angular/core';
 import { NgbActiveModal } from '@ng-bootstrap/ng-bootstrap';
 
 @Component({
-  selector: 'app-quit-game-confirm',
+  selector: 'app-missing-level-selection-modal',
   template: `
     <div class="modal-content">
       <div class="modal-header ">
-        <h4 class="modal-title w-100 text-center">Quitter</h4>
+        <h4 class="modal-title w-100 text-center"></h4>
         <button type="button" class="close" aria-label="Close" (click)="dismiss()">
           <span aria-hidden="true">&times;</span>
         </button>
       </div>
       <div class="modal-body w-100 text-center">
-        <p>Êtes vous sûr de vouloir quitter la partie en cours ?</p>
+        <p>Veuillez seléctionner une difficulté</p>
       </div>
       <div class="modal-footer justify-content-center">
-        <button type="button" class="btn-yes" (click)="close()">Oui</button>
-        <button type="button" class="btn-no" (click)="dismiss()">Non</button>
+        <button type="button" class="btn-ok" (click)="close()">Ok</button>
       </div>
     </div>
   `,
   styleUrls: ['../../shared/modals.scss'],
 })
-export class QuitGameConfirmationModalComponent {
+export class MissingLevelSelectionModalComponent {
   constructor(private modal: NgbActiveModal) {}
 
   dismiss() {
