+++ src/app/modules/pregame/pregame.module.ts
@@ -6,9 +6,19 @@ import { PlayerSelectionComponent } from 'src/app/modules/pregame/player-selecti
 import { PregameRoutingModule } from 'src/app/modules/pregame/pregame-routing.module';
 import { PregameComponent } from 'src/app/modules/pregame/pregame.component';
 import { SharedModule } from 'src/app/modules/shared/shared.module';
+import { MissingPlayerNameModalComponent } from './missing-player-name-modal/missing-player-name-modal.component';
+import { MissingLevelSelectionModalComponent } from './missing-level-selection-modal/missing-level-selection-modal.component';
 
 @NgModule({
-  declarations: [PregameComponent, PlayerSelectionComponent, LevelSelectionComponent],
+  declarations: [
+    PregameComponent,
+    PlayerSelectionComponent,
+    LevelSelectionComponent,
+    MissingPlayerNameModalComponent,
+    MissingLevelSelectionModalComponent,
+  ],
+  entryComponents: [MissingPlayerNameModalComponent, MissingLevelSelectionModalComponent],
   imports: [CommonModule, FormsModule, SharedModule, PregameRoutingModule],
+  exports: [MissingPlayerNameModalComponent, MissingLevelSelectionModalComponent],
 })
 export class PregameModule {}
